+++ packages/react-compass/src/card/card-image.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import {StyledComponentProps} from '../utils/stitches.types'
 import {useDOMRef} from '../utils/use-dom-ref'
-import {StyledCardImage} from './cardImage.styles'
+import {StyledCardImage} from './card-image.styles'
 
 type Props = StyledComponentProps
